refactor: scrape.go by putting cache logic in another file
Pushed the cache logic in another file, so that scrape.go is smaller.
