+++ scrape/scrape.go
@@ -260,7 +260,7 @@ type scrapeLoopOptions struct {
 	interval        time.Duration
 	timeout         time.Duration
 	mrc             []*relabel.Config
-	cache           *scrapeCache
+	cache           *Cache
 }
 
 const maxAheadTime = 10 * time.Minute
@@ -297,7 +297,7 @@ func newScrapePool(cfg *config.ScrapeConfig, app storage.Appendable, jitterSeed 
 		// Update the targets retrieval function for metadata to a new scrape cache.
 		cache := opts.cache
 		if cache == nil {
-			cache = newScrapeCache()
+			cache = NewScrapeCache()
 		}
 		opts.target.SetMetadataStore(cache)
 
@@ -422,12 +422,12 @@ func (sp *scrapePool) reload(cfg *config.ScrapeConfig) error {
 
 	forcedErr := sp.refreshTargetLimitErr()
 	for fp, oldLoop := range sp.loops {
-		var cache *scrapeCache
+		var cache *Cache
 		if oc := oldLoop.getCache(); reuseCache && oc != nil {
 			oldLoop.disableEndOfRunStalenessMarkers()
 			cache = oc
 		} else {
-			cache = newScrapeCache()
+			cache = NewScrapeCache()
 		}
 
 		t := sp.activeTargets[fp]
@@ -836,21 +836,14 @@ type loop interface {
 	run(errc chan<- error)
 	setForcedError(err error)
 	stop()
-	getCache() *scrapeCache
+	getCache() *Cache
 	disableEndOfRunStalenessMarkers()
-}
-
-type cacheEntry struct {
-	ref      storage.SeriesRef
-	lastIter uint64
-	hash     uint64
-	lset     labels.Labels
 }
 
 type scrapeLoop struct {
 	scraper         scraper
 	l               log.Logger
-	cache           *scrapeCache
+	cache           *Cache
 	lastScrapeSize  int
 	buffers         *pool.Pool
 	jitterSeed      uint64
@@ -878,255 +871,6 @@ type scrapeLoop struct {
 	appendMetadataToWAL bool
 }
 
-// scrapeCache tracks mappings of exposed metric strings to label sets and
-// storage references. Additionally, it tracks staleness of series between
-// scrapes.
-type scrapeCache struct {
-	iter uint64 // Current scrape iteration.
-
-	// How many series and metadata entries there were at the last success.
-	successfulCount int
-
-	// Parsed string to an entry with information about the actual label set
-	// and its storage reference.
-	series map[string]*cacheEntry
-
-	// Cache of dropped metric strings and their iteration. The iteration must
-	// be a pointer so we can update it without setting a new entry with an unsafe
-	// string in addDropped().
-	droppedSeries map[string]*uint64
-
-	// seriesCur and seriesPrev store the labels of series that were seen
-	// in the current and previous scrape.
-	// We hold two maps and swap them out to save allocations.
-	seriesCur  map[uint64]labels.Labels
-	seriesPrev map[uint64]labels.Labels
-
-	metaMtx  sync.Mutex
-	metadata map[string]*metaEntry
-}
-
-// metaEntry holds meta information about a metric.
-type metaEntry struct {
-	metadata.Metadata
-
-	lastIter       uint64 // Last scrape iteration the entry was observed at.
-	lastIterChange uint64 // Last scrape iteration the entry was changed at.
-}
-
-func (m *metaEntry) size() int {
-	// The attribute lastIter although part of the struct it is not metadata.
-	return len(m.Help) + len(m.Unit) + len(m.Type)
-}
-
-func newScrapeCache() *scrapeCache {
-	return &scrapeCache{
-		series:        map[string]*cacheEntry{},
-		droppedSeries: map[string]*uint64{},
-		seriesCur:     map[uint64]labels.Labels{},
-		seriesPrev:    map[uint64]labels.Labels{},
-		metadata:      map[string]*metaEntry{},
-	}
-}
-
-func (c *scrapeCache) iterDone(flushCache bool) {
-	c.metaMtx.Lock()
-	count := len(c.series) + len(c.droppedSeries) + len(c.metadata)
-	c.metaMtx.Unlock()
-
-	if flushCache {
-		c.successfulCount = count
-	} else if count > c.successfulCount*2+1000 {
-		// If a target had varying labels in scrapes that ultimately failed,
-		// the caches would grow indefinitely. Force a flush when this happens.
-		// We use the heuristic that this is a doubling of the cache size
-		// since the last scrape, and allow an additional 1000 in case
-		// initial scrapes all fail.
-		flushCache = true
-		targetScrapeCacheFlushForced.Inc()
-	}
-
-	if flushCache {
-		// All caches may grow over time through series churn
-		// or multiple string representations of the same metric. Clean up entries
-		// that haven't appeared in the last scrape.
-		for s, e := range c.series {
-			if c.iter != e.lastIter {
-				delete(c.series, s)
-			}
-		}
-		for s, iter := range c.droppedSeries {
-			if c.iter != *iter {
-				delete(c.droppedSeries, s)
-			}
-		}
-		c.metaMtx.Lock()
-		for m, e := range c.metadata {
-			// Keep metadata around for 10 scrapes after its metric disappeared.
-			if c.iter-e.lastIter > 10 {
-				delete(c.metadata, m)
-			}
-		}
-		c.metaMtx.Unlock()
-
-		c.iter++
-	}
-
-	// Swap current and previous series.
-	c.seriesPrev, c.seriesCur = c.seriesCur, c.seriesPrev
-
-	// We have to delete every single key in the map.
-	for k := range c.seriesCur {
-		delete(c.seriesCur, k)
-	}
-}
-
-func (c *scrapeCache) get(met string) (*cacheEntry, bool) {
-	e, ok := c.series[met]
-	if !ok {
-		return nil, false
-	}
-	e.lastIter = c.iter
-	return e, true
-}
-
-func (c *scrapeCache) addRef(met string, ref storage.SeriesRef, lset labels.Labels, hash uint64) {
-	if ref == 0 {
-		return
-	}
-	c.series[met] = &cacheEntry{ref: ref, lastIter: c.iter, lset: lset, hash: hash}
-}
-
-func (c *scrapeCache) addDropped(met string) {
-	iter := c.iter
-	c.droppedSeries[met] = &iter
-}
-
-func (c *scrapeCache) getDropped(met string) bool {
-	iterp, ok := c.droppedSeries[met]
-	if ok {
-		*iterp = c.iter
-	}
-	return ok
-}
-
-func (c *scrapeCache) trackStaleness(hash uint64, lset labels.Labels) {
-	c.seriesCur[hash] = lset
-}
-
-func (c *scrapeCache) forEachStale(f func(labels.Labels) bool) {
-	for h, lset := range c.seriesPrev {
-		if _, ok := c.seriesCur[h]; !ok {
-			if !f(lset) {
-				break
-			}
-		}
-	}
-}
-
-func (c *scrapeCache) setType(metric []byte, t textparse.MetricType) {
-	c.metaMtx.Lock()
-
-	e, ok := c.metadata[yoloString(metric)]
-	if !ok {
-		e = &metaEntry{Metadata: metadata.Metadata{Type: textparse.MetricTypeUnknown}}
-		c.metadata[string(metric)] = e
-	}
-	if e.Type != t {
-		e.Type = t
-		e.lastIterChange = c.iter
-	}
-	e.lastIter = c.iter
-
-	c.metaMtx.Unlock()
-}
-
-func (c *scrapeCache) setHelp(metric, help []byte) {
-	c.metaMtx.Lock()
-
-	e, ok := c.metadata[yoloString(metric)]
-	if !ok {
-		e = &metaEntry{Metadata: metadata.Metadata{Type: textparse.MetricTypeUnknown}}
-		c.metadata[string(metric)] = e
-	}
-	if e.Help != yoloString(help) {
-		e.Help = string(help)
-		e.lastIterChange = c.iter
-	}
-	e.lastIter = c.iter
-
-	c.metaMtx.Unlock()
-}
-
-func (c *scrapeCache) setUnit(metric, unit []byte) {
-	c.metaMtx.Lock()
-
-	e, ok := c.metadata[yoloString(metric)]
-	if !ok {
-		e = &metaEntry{Metadata: metadata.Metadata{Type: textparse.MetricTypeUnknown}}
-		c.metadata[string(metric)] = e
-	}
-	if e.Unit != yoloString(unit) {
-		e.Unit = string(unit)
-		e.lastIterChange = c.iter
-	}
-	e.lastIter = c.iter
-
-	c.metaMtx.Unlock()
-}
-
-func (c *scrapeCache) GetMetadata(metric string) (MetricMetadata, bool) {
-	c.metaMtx.Lock()
-	defer c.metaMtx.Unlock()
-
-	m, ok := c.metadata[metric]
-	if !ok {
-		return MetricMetadata{}, false
-	}
-	return MetricMetadata{
-		Metric: metric,
-		Type:   m.Type,
-		Help:   m.Help,
-		Unit:   m.Unit,
-	}, true
-}
-
-func (c *scrapeCache) ListMetadata() []MetricMetadata {
-	c.metaMtx.Lock()
-	defer c.metaMtx.Unlock()
-
-	res := make([]MetricMetadata, 0, len(c.metadata))
-
-	for m, e := range c.metadata {
-		res = append(res, MetricMetadata{
-			Metric: m,
-			Type:   e.Type,
-			Help:   e.Help,
-			Unit:   e.Unit,
-		})
-	}
-	return res
-}
-
-// MetadataSize returns the size of the metadata cache.
-func (c *scrapeCache) SizeMetadata() (s int) {
-	c.metaMtx.Lock()
-	defer c.metaMtx.Unlock()
-	for _, e := range c.metadata {
-		s += e.size()
-	}
-
-	return s
-}
-
-// MetadataLen returns the number of metadata entries in the cache.
-func (c *scrapeCache) LengthMetadata() int {
-	c.metaMtx.Lock()
-	defer c.metaMtx.Unlock()
-
-	return len(c.metadata)
-}
-
 func newScrapeLoop(ctx context.Context,
 	sc scraper,
 	l log.Logger,
@@ -1134,7 +878,7 @@ func newScrapeLoop(ctx context.Context,
 	sampleMutator labelsMutator,
 	reportSampleMutator labelsMutator,
 	appender func(ctx context.Context) storage.Appender,
-	cache *scrapeCache,
+	cache *Cache,
 	jitterSeed uint64,
 	honorTimestamps bool,
 	sampleLimit int,
@@ -1153,7 +897,7 @@ func newScrapeLoop(ctx context.Context,
 		buffers = pool.New(1e3, 1e6, 3, func(sz int) interface{} { return make([]byte, 0, sz) })
 	}
 	if cache == nil {
-		cache = newScrapeCache()
+		cache = NewScrapeCache()
 	}
 
 	appenderCtx := ctx
@@ -1443,7 +1187,7 @@ func (sl *scrapeLoop) disableEndOfRunStalenessMarkers() {
 	sl.disabledEndOfRunStalenessMarkers = true
 }
 
-func (sl *scrapeLoop) getCache() *scrapeCache {
+func (sl *scrapeLoop) getCache() *Cache {
 	return sl.cache
 }
 
@@ -1484,7 +1228,7 @@ func (sl *scrapeLoop) append(app storage.Appender, b []byte, contentType string,
 
 		sl.cache.metaMtx.Lock()
 		defer sl.cache.metaMtx.Unlock()
-		metaEntry, metaOk := sl.cache.metadata[yoloString([]byte(lset.Get(labels.MetricName)))]
+		metaEntry, metaOk := sl.cache.metadata[YoloString([]byte(lset.Get(labels.MetricName)))]
 		if metaOk && (isNewSeries || metaEntry.lastIterChange == sl.cache.iter) {
 			metadataChanged = true
 			meta.Type = metaEntry.Type
@@ -1548,10 +1292,10 @@ loop:
 		meta = metadata.Metadata{}
 		metadataChanged = false
 
-		if sl.cache.getDropped(yoloString(met)) {
+		if sl.cache.getDropped(YoloString(met)) {
 			continue
 		}
-		ce, ok := sl.cache.get(yoloString(met))
+		ce, ok := sl.cache.get(YoloString(met))
 		var (
 			ref  storage.SeriesRef
 			lset labels.Labels
@@ -1673,7 +1417,7 @@ loop:
 	return
 }
 
-func yoloString(b []byte) string {
+func YoloString(b []byte) string {
 	return *((*string)(unsafe.Pointer(&b)))
 }
